requirements.txt
alembic
asyncpg
discord.py>=2.3.0
python-dotenv
SQLAlchemy
PyNaCl